chore: apply changes from #4238
